requirements.txt


fastapi
uvicorn
sqlalchemy
aiosqlite

pydantic[email]

bcrypt==4.0.1 # For password hashing
passlib[bcrypt]
python-jose[cryptography]
python-multipart


pytest
httpx
pytest-asyncio